# server/swagger.yaml
openapi: 3.0.3
info:
  title: MyLoadTest Datapool API
  description: The MyLoadTest Datapool is a centralised, persistent data store. ...TODO
  version: 0.0.2
  contact:
    name: MyLoadTest
    url: https://www.myloadtest.com
servers:
  - url: http://localhost:3000
    description: Self-hosted datapool
...
components:
  schemas:
    Counter:
      description: a counter value (must be a positive whole number)
      type: integer
      minimum: 0
      example: 99
    MapValue:
      # A schema without a type matches any data type (except null).
      # The use-case for storing an array is unclear, but there is no reason to exclude it.
      description: Can be any value - string, number, boolean, array or object.
      example:
        firstName: John
        lastName: Smith
        registered: true
...
paths:
...
  /counter/{name}:
    post:
      summary: Initialise a counter
      description: Initialise a counter with a custom starting value. If you just need a counter that starts
        at 1, you can call ```GET /counter/{name}``` directly.
      tags: [counter]
      parameters:
        - in: path
          name: name
          description: The name of the counter
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Counter'
      responses:
        201:
          description: An empty response body is returned when the counter is initialised or reset.
        400:
          description: Bad request. Counter value must be a positive integer.
    get:
      summary: Get the value of a counter
      description: Get the value of a counter (counter names are case sensitive).
      tags: [counter]
      parameters:
        - in: path
          name: name
          description: The name of the counter
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returns the counter value. Note that a 404 response is not returned if the counter does not exist. Instead, the counter is created and initialised to 1.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Counter'
        400:
          description: The counter name is invalid
    delete:
      summary: Remove or reset a counter
      description: Remove or reset a counter (no option to delete multiple counters at the same time)
      tags: [counter]
      parameters:
        - in: path
          name: name
          description: The name of the counter
          required: true
          schema:
            type: string
      responses:
        200:
          description: The counter was deleted successfully. Note that there is no point returning a 404 if the counter is not found, as it is the same outcome if it is deleted or never existed in the first place.
        400:
          description: The counter name is invalid
...
  /map/{name}:
    post:
      summary: Save a value to a key
      description: Save a value to a map key. If the key already exists, the value will be overwritten. The value may be a JSON object, array or primitive value. If you need to avoid overwriting existing keys then check whether the key exists with ```GET /map/{name}``` first.
      tags: [map]
      parameters:
        - in: path
          name: name
          description: The name of the map key
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MapValue'
      responses:
        201:
          description: An empty response body is returned when the value is saved.
        400:
          description: Bad request. Saved value must be a valid object or primitive value.
    get:
      summary: Get the value associated with a key
      description: Get the value for a key-value pair (map keys are case sensitive).
      tags: [map]
      parameters:
        - in: path
          name: name
          description: The name of the map key
          required: true
          schema:
            type: string
      responses:
        200:
          description: A JSON object or primitive value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MapValue'
        400:
          description: The map name is invalid
        404:
          description: The map name was not found
    delete:
      summary: Remove a key-value pair
      description: Remove a map (no option to delete multiple maps at the same time)
      tags: [map]
      parameters:
        - in: path
          name: name
          description: The name of the map key
          required: true
          schema:
            type: string
      responses:
        200:
          description: The map was deleted successfully. Note that there is no point returning a 404 if the map key is not found, as it is the same outcome if it is deleted or never existed in the first place.
        400:
          description: The map name is invalid
...